+++ src/controller/python/chip/clusters/CHIPClusters.cpp
@@ -1403,8 +1403,9 @@ chip::ChipError::StorageType chip_ime_AppendCommand_ColorControl_StepColorTemper
 }
 chip::ChipError::StorageType chip_ime_AppendCommand_ColorControl_StepHue(chip::Controller::Device * device,
                                                                          chip::EndpointId ZCLendpointId, chip::GroupId,
-                                                                         uint8_t stepMode, uint8_t stepSize, uint8_t transitionTime,
-                                                                         uint8_t optionsMask, uint8_t optionsOverride)
+                                                                         uint8_t stepMode, uint8_t stepSize,
+                                                                         uint16_t transitionTime, uint8_t optionsMask,
+                                                                         uint8_t optionsOverride)
 {
     VerifyOrReturnError(device != nullptr, CHIP_ERROR_INVALID_ARGUMENT.AsInteger());
     chip::Controller::ColorControlCluster cluster;
@@ -1414,7 +1415,7 @@ chip::ChipError::StorageType chip_ime_AppendCommand_ColorControl_StepHue(chip::C
 chip::ChipError::StorageType chip_ime_AppendCommand_ColorControl_StepSaturation(chip::Controller::Device * device,
                                                                                 chip::EndpointId ZCLendpointId, chip::GroupId,
                                                                                 uint8_t stepMode, uint8_t stepSize,
-                                                                                uint8_t transitionTime, uint8_t optionsMask,
+                                                                                uint16_t transitionTime, uint8_t optionsMask,
                                                                                 uint8_t optionsOverride)
 {
     VerifyOrReturnError(device != nullptr, CHIP_ERROR_INVALID_ARGUMENT.AsInteger());
